india_news_telegram_crawler/requirements.txt
requests>=2.25.0
beautifulsoup4>=4.9.3
python-dateutil>=2.8.1
pandas>=1.1.5
openpyxl>=3.0.5
python-telegram-bot>=13.7
python-dotenv>=0.19.0
